requirements.txt
langchain==0.2.1
faiss-cpu
sentence-transformers
python-dotenv
httpx
openai
huggingface-hub